Rename sort.ts to sort-embedded.ts
